Move memory-debugger and pool-monitor into a debug directory
