+++ lib/processes/js-min-child-process-pool.js
@@ -2,7 +2,7 @@
 const { fork } = require( 'child_process' );
 const { createPool } = require( 'generic-pool' );
 const path = require( 'path' );
-const maybeEnablePoolMonitor = require( '../pool-monitor' );
+const maybeEnablePoolMonitor = require( '../debug/pool-monitor' );
 
 const factory = {
 	create: function () {
